Rebuild the UI
This change rebuilds the core UI to have a more traditional media player
like interface on a second page of a view pager.

In order to keep the logic somewhat simple for this change there's not a
page indicator, nor any restrictions on flipping from the media input
(media id, search, uri, or a simple 'play') to the status screen.

Once this is merged I'll change it so there must be some media playing
to switch to the second screen, and indeed switch automatically once the
developer presses 'play' or 'prepare'.

Change-Id: I9b5d367497ddfee939e36b0880f4e8e95e576194