Add a new way to pass data into the app via adb.
Originally there was only a set of extras that could be specified on the
command line to set the package name, media browser component, and pass
in a URI, search, or media ID.

I've updated it to allow passing in the package name and a query
that the developer may want to execute, and I've done it via a 'deep
link' type of structure. I hope this will make it relatively easy for
developers to use, and won't require using very short names for extras,
which is typically not recommended best practice.

If developers require the ability to pass in complex URIs (with
multiple parameters) then they can fall back to an updated set of extras
that follow recommended best practices in their naming.

Change-Id: Ic6df65e9350d430e7fd9204ad4c458b117b1a300