Add support for shuffle and repeat "modes"
setShuffleEnabled is deprecated in favour of "modes" that allow more
control over the shuffle style (e.g. shuffle all, 'group', or none)
instead of just shuffle or not.

Also added support for repeat modes (one, group, all), and updated the
controls UI to display the mode as well as to cycle between the modes
instead of just toggling shuffle on or off.