Fix crash when attempting to connect via an Intent.
A race condition was introduced during refactor where the UI attempted
to update itself before a connection was made to the MediaSession when
connection information (by way of package name) was passed in via an
Intent.

The change makes it so the ViewPager is marked as GONE if the connection
hasn't yet been established. After a connection is then established, the
view is set to VISIBLE to trigger the updates.

Change-Id: Ie713a70df7474786688f286cbe0e49eaa63a971e