Configure steps should populate origState/Metadata
This is necessary for chained tests (e.g. if from STATE_STOPPED, a
Play-Pause test is run, the Play half of the test would mark
STATE_STOPPED as valid since STATE_STOPPED is the original state for
that half of the test, but the Pause half should not consider
STATE_STOPPED to be valid since the Pause half starts in STATE_PLAYING).

Change-Id: I9aa16f1b4c1a60e47fe805f933fcbdd933aed396